+++ src/shapes/polyline/mod.rs
@@ -39,9 +39,10 @@ impl Segment {
     }
 }
 
+#[derive(Debug, Clone, PartialEq)]
 pub struct Polyline {
     segments: Vec<Segment>,
-    pub color: Color,
+    color: Color,
     closed: bool,
 }
 
@@ -63,13 +64,14 @@ impl Polyline {
     }
 
     pub fn rounded_rect(
-        left: usize,
-        top: usize,
-        right: usize,
-        bottom: usize,
+        left: isize,
+        top: isize,
+        right: isize,
+        bottom: isize,
         corner_size: usize,
         color: Color,
     ) -> Result<Self, PolylineError> {
+        let corner_size = corner_size as isize;
         let tl_arc = Coord::from((left + corner_size, top + corner_size));
         let tr_arc = Coord::from((right - corner_size, top + corner_size));
         let bl_arc = Coord::from((left + corner_size, bottom - corner_size));
@@ -82,13 +84,21 @@ impl Polyline {
 
         Polyline::start((left + corner_size, top), color)
             .add_line_to(line1_end)?
-            .add_arc_around(tr_arc, corner_size, 0, 90)?
+            .add_arc_around(tr_arc, corner_size as usize, 0, 90)?
             .add_line_to(line2_end)?
-            .add_arc_around(br_arc, corner_size, 90, 90)?
+            .add_arc_around(br_arc, corner_size as usize, 90, 90)?
             .add_line_to(line3_end)?
-            .add_arc_around(bl_arc, corner_size, 180, 90)?
+            .add_arc_around(bl_arc, corner_size as usize, 180, 90)?
             .add_line_to(line4_end)?
-            .add_arc_around(tl_arc, corner_size, 270, 90)
+            .add_arc_around(tl_arc, corner_size as usize, 270, 90)
+    }
+}
+
+impl Polyline {
+    pub fn with_color(&self, color: Color) -> Self {
+        let mut cloned = self.clone();
+        cloned.color = color;
+        cloned
     }
 }
 
@@ -172,7 +182,7 @@ mod test {
                 LineTo(Coord::new(20, 10)),
                 LineTo(Coord::new(20, 20)),
                 LineTo(Coord::new(10, 20)),
-                LineTo(Coord::new(10, 10))
+                LineTo(Coord::new(10, 10)),
             ]
         )
     }